[api] refactoring: start basic class for discrete concept needing codegen
