+++ spl/api/__init__.py
@@ -1,5 +1,6 @@
 # -*- coding: UTF-8 -*-
 
+from spl.api import basic
 from spl.api import discretization
 from spl.api import essential_bc
 from spl.api import ast
